Add new custom command one-time dialog
