+++ Tool/Sources/Preferences/Types/CustomCommand.swift
@@ -1,27 +1,46 @@
-import Foundation
 import CryptoKit
+import Foundation
 
 public struct CustomCommand: Codable, Equatable {
     /// The custom command feature.
     ///
     /// Keep everything optional so nothing will break when the format changes.
     public enum Feature: Codable, Equatable {
-        case promptToCode(extraSystemPrompt: String?, prompt: String?, continuousMode: Bool?, generateDescription: Bool?)
-        case chatWithSelection(extraSystemPrompt: String?, prompt: String?, useExtraSystemPrompt: Bool?)
+        /// Prompt to code.
+        case promptToCode(
+            extraSystemPrompt: String?,
+            prompt: String?,
+            continuousMode: Bool?,
+            generateDescription: Bool?
+        )
+        /// Send message.
+        case chatWithSelection(
+            extraSystemPrompt: String?,
+            prompt: String?,
+            useExtraSystemPrompt: Bool?
+        )
+        /// Custom chat.
         case customChat(systemPrompt: String?, prompt: String?)
+        /// One-time dialog.
+        case oneTimeDialog(
+            systemPrompt: String?,
+            overwriteSystemPrompt: Bool?,
+            prompt: String?,
+            receiveReplyInNotification: Bool?
+        )
     }
 
     public var id: String { commandId ?? legacyId }
     public var commandId: String?
     public var name: String
     public var feature: Feature
-    
+
     public init(commandId: String, name: String, feature: Feature) {
         self.commandId = commandId
         self.name = name
         self.feature = feature
     }
-    
+
     var legacyId: String {
         name.sha1HexString
     }
@@ -41,3 +60,4 @@ private extension String {
         Insecure.SHA1.hash(data: data(using: .utf8) ?? Data()).hexStr
     }
 }
+
